+++ doc/whats-new.rst
@@ -44,6 +44,9 @@ New Features
   By `Thomas Hirtz <https://github.com/thomashirtz>`_.
 - allow passing a function to ``combine_attrs`` (:pull:`4896`).
   By `Justus Magin <https://github.com/keewis>`_.
+- The values stored in an unnamed dataarray can now be referenced in a call to
+  py:func:`xarray.DataArray.query` via 'self' (:issue:`5492`, :pull:`5493`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~